refactor: :recycle: Changed create statement of the posts table.
Laravel by default works with created_at and updated_at fields. Had to change date field to created_at and added a new updated_at field that is going to be useful when implementing the edit post feature. Also had to change the types of the fields to TIMESTAMP since laravel works with TIMESTAMP type by default and not DATE type.