fix: :bug: Comment likes implemented.
This feature was implemented by Filipe. When I was merging his branch to main I deleted his branch and when I checked the content of the main branch there were conflicts so I had to revert the merge. However I lost his branch so I had to copy the code he wrote to main.